refactor config module
- split config module to mods.rs and metadata.rs
- move file_info.rs to config
